Move nl/nl to nl/ext
