+++ src/theory/arith/nl/ext/factoring_check.h
@@ -12,8 +12,8 @@
  ** \brief Check for factoring lemma
  **/
 
-#ifndef CVC4__THEORY__ARITH__NL__NL__FACTORING_CHECK_H
-#define CVC4__THEORY__ARITH__NL__NL__FACTORING_CHECK_H
+#ifndef CVC4__THEORY__ARITH__NL__EXT__FACTORING_CHECK_H
+#define CVC4__THEORY__ARITH__NL__EXT__FACTORING_CHECK_H
 
 #include <vector>
