+++ src/theory/arith/nl/ext/monomial.cpp
@@ -1,5 +1,5 @@
 /*********************                                                        */
-/*! \file nl_monomial.cpp
+/*! \file monomial.cpp
  ** \verbatim
  ** Top contributors (to current version):
  **   Andrew Reynolds, Tim King, Andres Noetzli
@@ -12,7 +12,7 @@
  ** \brief Implementation of utilities for monomials
  **/
 
-#include "theory/arith/nl/nl/monomial.h"
+#include "theory/arith/nl/ext/monomial.h"
 
 #include "theory/arith/arith_utilities.h"
 #include "theory/arith/nl/nl_lemma_utils.h"
