+++ src/theory/arith/nl/ext/monomial_bounds_check.h
@@ -12,12 +12,12 @@
  ** \brief Check for monomial bound inference lemmas
  **/
 
-#ifndef CVC4__THEORY__ARITH__NL__NL__MONOMIAL_BOUNDS_CHECK_H
-#define CVC4__THEORY__ARITH__NL__NL__MONOMIAL_BOUNDS_CHECK_H
+#ifndef CVC4__THEORY__ARITH__NL__EXT__MONOMIAL_BOUNDS_CHECK_H
+#define CVC4__THEORY__ARITH__NL__EXT__MONOMIAL_BOUNDS_CHECK_H
 
 #include "expr/node.h"
-#include "theory/arith/nl/nl/constraint.h"
-#include "theory/arith/nl/nl/shared_check_data.h"
+#include "theory/arith/nl/ext/constraint.h"
+#include "theory/arith/nl/ext/shared_check_data.h"
 
 namespace CVC4 {
 namespace theory {
