+++ src/theory/arith/nl/ext/split_zero_check.h
@@ -12,11 +12,11 @@
  ** \brief Check for split zero lemma
  **/
 
-#ifndef CVC4__THEORY__ARITH__NL__NL__SPLIT_ZERO_CHECK_H
-#define CVC4__THEORY__ARITH__NL__NL__SPLIT_ZERO_CHECK_H
+#ifndef CVC4__THEORY__ARITH__NL__EXT__SPLIT_ZERO_CHECK_H
+#define CVC4__THEORY__ARITH__NL__EXT__SPLIT_ZERO_CHECK_H
 
 #include "expr/node.h"
-#include "theory/arith/nl/nl/shared_check_data.h"
+#include "theory/arith/nl/ext/shared_check_data.h"
 
 namespace CVC4 {
 namespace theory {
