+++ src/theory/arith/nl/nonlinear_extension.h
@@ -28,14 +28,14 @@
 #include "expr/node.h"
 #include "theory/arith/inference_manager.h"
 #include "theory/arith/nl/cad_solver.h"
+#include "theory/arith/nl/ext/factoring_check.h"
+#include "theory/arith/nl/ext/monomial_bounds_check.h"
+#include "theory/arith/nl/ext/monomial_check.h"
+#include "theory/arith/nl/ext/split_zero_check.h"
+#include "theory/arith/nl/ext/tangent_plane_check.h"
 #include "theory/arith/nl/ext_theory_callback.h"
 #include "theory/arith/nl/iand_solver.h"
 #include "theory/arith/nl/icp/icp_solver.h"
-#include "theory/arith/nl/nl/factoring_check.h"
-#include "theory/arith/nl/nl/monomial_bounds_check.h"
-#include "theory/arith/nl/nl/monomial_check.h"
-#include "theory/arith/nl/nl/split_zero_check.h"
-#include "theory/arith/nl/nl/tangent_plane_check.h"
 #include "theory/arith/nl/nl_lemma_utils.h"
 #include "theory/arith/nl/nl_model.h"
 #include "theory/arith/nl/stats.h"
